shader defines refactor - part #1

+++ src/renderer/renderer.h
@@ -33,7 +33,7 @@ class LUMIX_RENDERER_API Renderer : public IPlugin
 		virtual void viewCounterAdd() = 0;
 		virtual void makeScreenshot(const Path& filename) = 0;
 		virtual int getPassIdx(const char* pass) = 0;
-		virtual int getShaderDefineIdx(const char* define) = 0;
+		virtual uint8 getShaderDefineIdx(const char* define) = 0;
 		virtual const char* getShaderDefine(int define_idx) = 0;
 		virtual LIFOAllocator& getFrameAllocator() = 0;
 		virtual const bgfx::VertexDecl& getBasicVertexDecl() const = 0;
